Add analysis of hadoop token files and warns if they are over 12 old, or just missing.
This is to debug situations where out of date DTs are blocking auth via core-site values.

+++ src/main/java/org/apache/hadoop/fs/store/commands/LocateFiles.java
@@ -122,7 +122,7 @@ public class LocateFiles extends StoreEntryPoint {
     long files = count.get();
     double millisPerFile = files > 0 ? (((float)duration.value()) / files) : 0;
     long totalSize = size.get();
-    long bytesPerFile = (long) (files > 0 ? totalSize / files : 0);
+    long bytesPerFile = (files > 0 ? totalSize / files : 0);
     println("");
     println("Found %s files, %,.0f milliseconds per file",
         files, millisPerFile);
